pm: submodules for cortexm_common and cleanup of pm_layered
