+++ cpu/cc2538/Makefile.include
@@ -2,6 +2,8 @@ export CPU_ARCH := cortex-m3
 
 # include common SPI functions
 USEMODULE += periph_common
-USEMODULE += pm_cortexm_common
+USEMODULE += pm_cortexm_common_set_lowest
+USEMODULE += pm_cortexm_common_off
+USEMODULE += pm_cortexm_common_reboot
 
 include $(RIOTMAKE)/arch/cortexm.inc.mk
