+++ cpu/cc26x0/Makefile.include
@@ -1,5 +1,7 @@
 export CPU_ARCH := cortex-m3
 
-USEMODULE += pm_cortexm_common
+USEMODULE += pm_cortexm_common_set_lowest
+USEMODULE += pm_cortexm_common_off
+USEMODULE += pm_cortexm_common_reboot
 
 include $(RIOTMAKE)/arch/cortexm.inc.mk
