+++ cpu/cortexm_common/pm/Makefile
@@ -1,3 +1,3 @@
-MODULE = pm_cortexm_common
-
+MODULE := pm_cortexm_common
+SUBMODULES := 1
 include $(RIOTBASE)/Makefile.base
