+++ cpu/ezr32wg/Makefile.include
@@ -1,5 +1,7 @@
 export CPU_ARCH = cortex-m4f
 
-USEMODULE += pm_cortexm_common
+USEMODULE += pm_cortexm_common_set_lowest
+USEMODULE += pm_cortexm_common_off
+USEMODULE += pm_cortexm_common_reboot
 
 include $(RIOTMAKE)/arch/cortexm.inc.mk
