+++ cpu/lm4f120/Makefile.include
@@ -1,9 +1,10 @@
 export CPU_ARCH = cortex-m4f
 
-include $(RIOTMAKE)/arch/cortexm.inc.mk
-
 # use common periph functions
 USEMODULE += periph_common
-USEMODULE += pm_cortexm_common
+USEMODULE += pm_cortexm_common_set_lowest
+USEMODULE += pm_cortexm_common_off
+USEMODULE += pm_cortexm_common_reboot
 
+include $(RIOTMAKE)/arch/cortexm.inc.mk
 include $(RIOTCPU)/stellaris_common/Makefile.include
