+++ cpu/stm32_common/pm/pm.c
@@ -61,14 +61,3 @@ void pm_set(unsigned mode)
 
     cortexm_sleep(deep);
 }
-
-/*
-pm_layered uses pm_set
-#if defined(CPU_FAM_STM32F1) || defined(CPU_FAM_STM32F2) || defined(CPU_FAM_STM32F4)
-void pm_off(void)
-{
-    irq_disable();
-    pm_set_lowest();
-}
-#endif
-*/
