+++ cpu/stm32f0/Makefile.include
@@ -1,10 +1,11 @@
 export CPU_ARCH = cortex-m0
 export CPU_FAM  = stm32f0
 
-#USEMODULE += pm_cortexm_common
 # based on stm32 -> can we use pm_layered and pm_stm32_common_layered?
 USEMODULE += pm_layered
 USEMODULE += pm_stm32_common_layered
+USEMODULE += pm_cortexm_common_off
+USEMODULE += pm_cortexm_common_reboot
 
 include $(RIOTCPU)/stm32_common/Makefile.include
 include $(RIOTMAKE)/arch/cortexm.inc.mk
