+++ cpu/stm32f4/Makefile.include
@@ -3,6 +3,8 @@ export CPU_FAM  = stm32f4
 
 USEMODULE += pm_layered
 USEMODULE += pm_stm32_common_layered
+USEMODULE += pm_cortexm_common_off
+USEMODULE += pm_cortexm_common_reboot
 
 include $(RIOTCPU)/stm32_common/Makefile.include
 include $(RIOTMAKE)/arch/cortexm.inc.mk
