+++ cpu/stm32l1/Makefile.include
@@ -1,9 +1,10 @@
 export CPU_ARCH = cortex-m3
 export CPU_FAM  = stm32l1
 
-#USEMODULE += pm_cortexm_common
 USEMODULE += pm_layered
 USEMODULE += pm_stm32_common_layered
+USEMODULE += pm_cortexm_common_off
+USEMODULE += pm_cortexm_common_reboot
 
 include $(RIOTCPU)/stm32_common/Makefile.include
 include $(RIOTMAKE)/arch/cortexm.inc.mk
