+++ sys/pm_layered/pm.c
@@ -86,12 +86,3 @@ void pm_unblock(unsigned mode)
     pm_blocker.val_u8[mode]--;
     irq_restore(state);
 }
-
-void __attribute__((weak)) pm_off(void)
-{
-    pm_blocker.val_u32 = 0;
-    pm_set_lowest();
-    while(1);
-}
-
-void __attribute__((weak)) pm_reboot(void) { }
